Add (mostly) complete contribution documentation
Still some explanations missing, TODOs to take care of, inconsistencies
in formatting to resolve, etc, but this should be enough for any
beginner looking to make a contribution. Figured it's better to push
as-is rather than sit on it for another month or more (depending on how
long it takes me to satisfy my perfectionism). Besides, the more
uncommitted stuff that piles up, the less I end up wanting to work on
any of it.